backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
supabase
httpx
google-generativeai